Adding wait-for-it-listen.sh which listens instead of scanning ports
FYI, I am using this on an Alpine image without bash.
The busybox ash does NOT support arrays and has some other
compatibility issues, so fixed those.

A hacky way to make this work is to write 'host:', ':port'
instead of 'host:port'. If there is a cleaner way around
this I'd prefer it.

Have a use case where the other container is not ready even
if available at a port as I am doing a database initialization
from a dump file and this might take a while.

I have another shell script on the database container that scans
the port specified in this script once it is done so in essence
this reverses the functionality of wait-for-it.

Hopefully this is usefull to other people.